Made `aries.simulate` return a list of states.
Previously, the ARIES simulator would create a single ARIES state and
thread it through functions like `analysis`, `redo`, and `undo` that
simulated the execution of ARIES. These functions modified the state
object in place, so there was only ever one state object.

This commit makes the ARIES simulator instead return a list of states,
each of which is derived from the last. Having a list of states makes it
much easier for the front end to step forwards and backwards through the
execution of the simulation.

Changing the code to return a list of states was actually quite easy. I
simply deep copy the state objects after each step in the simulation.