Fixed analysis aborted transaction bug.
Previously, during the analysis phase, a transaction was marked aborted
when an update was found for it. A transaction was marked committed if a
commit was found for it. This is buggy because there may be an in
progress transaction in the transaction table that does not have an
update after the checkpoint. When analysis ends, the transaction's
status would be in progress, but it should be aborted.

Now, transaction status is changed to committed if a commit is found but
is otherwise left untouched until the end of the analysis phase at which
point all in progress transactions are marked aborted.